+++ src/data/projects.ts
@@ -1,0 +1,40 @@
+import { ProjectsType } from "./dataTypes";
+
+export const projects: Array<ProjectsType> = [
+  {
+    id: 1,
+    title: "The Shopping Booth",
+    techStacks: "React Js, Redux, Javascript, Shadcn/ui",
+    imgSrc: "webstorm.png",
+    altText: "The Shopping Booth",
+    github: "https://github.com/abhishekjaiswar221/the-shopping-booth",
+    live: "https://theshoppingbooth.vercel.app",
+  },
+  {
+    id: 2,
+    title: "Taskify Cloud Todo",
+    techStacks: "React Js, Javascript, Shadcn/ui, Tailwind",
+    imgSrc: "webstorm.png",
+    altText: "Taskify Cloud Todo",
+    github: "https://github.com/abhishekjaiswar221/taskify-cloud-todo",
+    live: "https://taskifytodo.vercel.app",
+  },
+  {
+    id: 3,
+    title: "Microsoft UI Clone",
+    techStacks: "React Js, Javascript, Tailwind",
+    imgSrc: "webstorm.png",
+    altText: "Microsoft UI Clone",
+    github: "https://github.com/abhishekjaiswar221/microsoft-ui-clone",
+    live: "https://microsoftui.vercel.app",
+  },
+  {
+    id: 4,
+    title: "News Monkey",
+    techStacks: "React Js, Javascript, Bootstrap",
+    imgSrc: "webstorm.png",
+    altText: "News Monkey",
+    github: "https://github.com/abhishekjaiswar221/news-monkey",
+    live: "https://microsoftui.vercel.app",
+  },
+];
