+++ src/data/socialIcons.ts
@@ -1,0 +1,31 @@
+import { SocialIconsType } from "./dataTypes";
+
+export const socialIcons: Array<SocialIconsType> = [
+  {
+    id: 1,
+    title: "Github",
+    imgSrc: "github.svg",
+    altText: "Github Logo",
+    bgColor: "bg-github",
+    borderColor: "hover:border-github",
+    href: "http://github.com/abhishekjaiswar221",
+  },
+  {
+    id: 2,
+    title: "LinkedIn",
+    imgSrc: "linkedin.svg",
+    altText: "LinkedIn Logo",
+    bgColor: "bg-linkedin",
+    borderColor: "hover:border-linkedin",
+    href: "https://linkedin.com/in/abhishekjai221",
+  },
+  {
+    id: 3,
+    title: "Twitter",
+    imgSrc: "twitter.svg",
+    altText: "Twitter Logo",
+    bgColor: "bg-twitter",
+    borderColor: "hover:border-twitter",
+    href: "https://x.com/abhishekjai221",
+  },
+];
