+++ src/data/techIcons.ts
@@ -1,0 +1,310 @@
+import { TechIconsType } from "./dataTypes";
+
+export const frontendTechIcons: Array<TechIconsType> = [
+  {
+    id: 1,
+    title: "HTML",
+    imgSrc: "html5.svg",
+    altText: "HTML Logo",
+  },
+  {
+    id: 2,
+    title: "CSS",
+    imgSrc: "css3.svg",
+    altText: "CSS Logo",
+  },
+  {
+    id: 3,
+    title: "Sass",
+    imgSrc: "sass.svg",
+    altText: "Sass Logo",
+  },
+  {
+    id: 4,
+    title: "Bootstrap",
+    imgSrc: "bootstrap.svg",
+    altText: "Bootstrap CSS Logo",
+  },
+  {
+    id: 5,
+    title: "Tailwind",
+    imgSrc: "tailwindcss.svg",
+    altText: "Tailwind CSS Logo",
+  },
+  {
+    id: 6,
+    title: "Shadcn/UI",
+    imgSrc: "shadcnui.svg",
+    altText: "Shadcn/ui Logo",
+  },
+  {
+    id: 7,
+    title: "Radix UI",
+    imgSrc: "radixui.svg",
+    altText: "Radix UI Logo",
+  },
+  // {
+  //   id: 8,
+  //   title: "Framer Motion",
+  //   imgSrc: "framer.svg",
+  //   altText: "Framer Motion Logo",
+  // },
+  {
+    id: 9,
+    title: "Javascript",
+    imgSrc: "js.svg",
+    altText: "Javascript Logo",
+  },
+  {
+    id: 10,
+    title: "React JS",
+    imgSrc: "reactjs.svg",
+    altText: "React JS Logo",
+  },
+  {
+    id: 11,
+    title: "Redux JS",
+    imgSrc: "redux.svg",
+    altText: "Redux JS Logo",
+  },
+  // {
+  //   id: 12,
+  //   title: "Zod",
+  //   imgSrc: "zod.svg",
+  //   altText: "Zod Logo",
+  // },
+];
+
+export const backend_devopsTechIcons: Array<TechIconsType> = [
+  {
+    id: 1,
+    title: "Java",
+    imgSrc: "java.svg",
+    altText: "Java Logo",
+  },
+  {
+    id: 2,
+    title: "MySQL",
+    imgSrc: "mysql.svg",
+    altText: "MySQL Logo",
+  },
+  {
+    id: 3,
+    title: "MongoDB",
+    imgSrc: "mongodb.svg",
+    altText: "MongoDB Logo",
+  },
+  {
+    id: 4,
+    title: "Git",
+    imgSrc: "git.svg",
+    altText: "Git Logo",
+  },
+  {
+    id: 5,
+    title: "Github",
+    imgSrc: "github.svg",
+    altText: "Github Logo",
+  },
+  {
+    id: 6,
+    title: "Vercel",
+    imgSrc: "vercel.svg",
+    altText: "Vercel Logo",
+  },
+  {
+    id: 7,
+    title: "Cloudflare",
+    imgSrc: "cloudflare.svg",
+    altText: "Cloudflare Logo",
+  },
+  {
+    id: 8,
+    title: "Netlify",
+    imgSrc: "netlify.svg",
+    altText: "Netlify Logo",
+  },
+  // {
+  //   id: 9,
+  //   title: "Mongoose Js",
+  //   imgSrc: "mongoose.svg",
+  //   altText: "Mongoose Logo",
+  // },
+];
+
+export const devtoolsTechIcons: Array<TechIconsType> = [
+  {
+    id: 1,
+    title: "VS Code",
+    imgSrc: "vscode.svg",
+    altText: "VS Code Logo",
+  },
+  {
+    id: 2,
+    title: "Webstorm",
+    imgSrc: "webstorm.svg",
+    altText: "Webstorm Logo",
+  },
+  {
+    id: 3,
+    title: "Intellij IDEA",
+    imgSrc: "intellij.svg",
+    altText: "Intellij IDEA Logo",
+  },
+  {
+    id: 4,
+    title: "Datagrip",
+    imgSrc: "datagrip.svg",
+    altText: "Datagrip Logo",
+  },
+  // {
+  //   id: 5,
+  //   title: "JetBrains",
+  //   imgSrc: "jetbrains.svg",
+  //   altText: "JetBrains Logo",
+  // },
+];
+
+export const othersTechIcons: Array<TechIconsType> = [
+  {
+    id: 1,
+    title: "REST API",
+    imgSrc: "restapi.svg",
+    altText: "REST API Logo",
+  },
+  {
+    id: 2,
+    title: "Postman",
+    imgSrc: "postman.svg",
+    altText: "Postman Logo",
+  },
+  {
+    id: 3,
+    title: "Vite JS",
+    imgSrc: "vitejs.svg",
+    altText: "Vite Js Logo",
+  },
+  {
+    id: 4,
+    title: "Rollup Js",
+    imgSrc: "rollup.svg",
+    altText: "Rollup Js Logo",
+  },
+  {
+    id: 5,
+    title: "Eslint",
+    imgSrc: "eslint.svg",
+    altText: "Eslint Logo",
+  },
+  {
+    id: 6,
+    title: "Canva",
+    imgSrc: "canva.svg",
+    altText: "Canva Logo",
+  },
+  // {
+  //   id: 7,
+  //   title: "Prettier",
+  //   imgSrc: "prettier.svg",
+  //   altText: "Prettier Logo",
+  // },
+  // {
+  //   id: 8,
+  //   title: "JSON",
+  //   imgSrc: "json.svg",
+  //   altText: "JSON Logo",
+  // },
+  // {
+  //   id: 9,
+  //   title: "Figma",
+  //   imgSrc: "figma.svg",
+  //   altText: "Figma Logo",
+  // },
+  // {
+  //   id: 10,
+  //   title: "Markdown",
+  //   imgSrc: "markdown.svg",
+  //   altText: "Markdown Logo",
+  // },
+];
+
+export const devopsTechIcons: Array<TechIconsType> = [
+  {
+    id: 1,
+    title: "Git",
+    imgSrc: "git.svg",
+    altText: "Git Logo",
+  },
+  {
+    id: 2,
+    title: "Github",
+    imgSrc: "github.svg",
+    altText: "Github Logo",
+  },
+  {
+    id: 3,
+    title: "Vercel",
+    imgSrc: "vercel.svg",
+    altText: "Vercel Logo",
+  },
+  {
+    id: 4,
+    title: "Cloudflare Pages",
+    imgSrc: "cloudflare.svg",
+    altText: "Cloudflare Pages Logo",
+  },
+  {
+    id: 5,
+    title: "Netlify",
+    imgSrc: "netlify.svg",
+    altText: "Netlify Logo",
+  },
+];
+
+export const build_lintingToolsTechIcons: Array<TechIconsType> = [
+  {
+    id: 1,
+    title: "Vite JS",
+    imgSrc: "vitejs.svg",
+    altText: "Vite Js Logo",
+  },
+  {
+    id: 2,
+    title: "Rollup Js",
+    imgSrc: "rollup.svg",
+    altText: "Rollup Js Logo",
+  },
+  {
+    id: 3,
+    title: "Eslint",
+    imgSrc: "eslint.svg",
+    altText: "Eslint Logo",
+  },
+  {
+    id: 4,
+    title: "Prettier",
+    imgSrc: "prettier.svg",
+    altText: "Prettier Logo",
+  },
+];
+
+export const packageManagersTechIcons: Array<TechIconsType> = [
+  {
+    id: 1,
+    title: "NPM",
+    imgSrc: "npm.svg",
+    altText: "NPM Logo",
+  },
+  {
+    id: 2,
+    title: "PNPM",
+    imgSrc: "pnpm.svg",
+    altText: "PNPM Logo",
+  },
+  {
+    id: 3,
+    title: "Yarn",
+    imgSrc: "yarn.svg",
+    altText: "Yarn Logo",
+  },
+];
